Avoid bare variable in when and the resulting warning.
Also, do not access attributes of a dict without ensuring that it is
defined. While the dict is now defined in defaults, let's be safe.

Using the default filter allows to entirely avoid the check if a
dictionary is defined: since Ansible 2.8, one can use the default filter
on attributes on an undefined dict, so the test is unnecessary.