support gather_facts: false; support setup-snapshot.yml
Some users use `gather_facts: false` in their playbooks.  This changes
the role to work in that case, by gathering only the facts it requires
to run.
CI testing can be sped up by creating a snapshot image pre-installed
with packages.  tests/setup-snapshot.yml can be used by a CI system
to do this.
Use `public: true` with the certificate role because the test uses
internal variables from that role.