fix: do not fail if authorized_keys not found
Cause: The evaluation of `__kdump_authorized_keys is defined` was happening
after the evaluation of `(__kdump_authorized_keys.content ...)`.  I guess
the parentheses cause the evaluation to happen first, regardless of the
`if` conditional.

Consequence: If authorized_keys were not found, the role would attempt to
evaluate the undefined `__kdump_authorized_keys` and fail.

Fix: Use the `defined` filter `d` with `__kdump_authorized_keys` to define
the value as an empty line, and omit the `if` clause.

Result: The role does not fail if authorized_keys are missing.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>