fix: Write new authorized_keys if needed is not idempotent
Cause: The new key content had an extra newline at the end.

Consequence: The test to see if the new key content was in the
current authorized_key list failed, so the key was added every
time, and the task was not idempotent.

Fix: Ensure the new key value has no extra newline. Use a simple
list `in` test to see if the new value is in the existing list.

Result: The task to write authorized_keys is idempotent.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>